requirements.txt
anyio==4.0.0
asgiref==3.7.2
certifi==2023.7.22
charset-normalizer==3.3.0
colorama==0.4.6
crispy-bootstrap5==0.7
Django==4.2.5
django-crispy-forms==2.0
django-jazzmin==2.6.0
django-restframework==0.0.1
djangorestframework==3.14.0
h11==0.14.0
httpcore==0.18.0
httpx==0.25.0
idna==3.4
Pillow==10.0.1
psycopg2==2.9.7
pypng==0.20220715.0
python-telegram-bot==20.6
pytz==2023.3.post1
qrcode==7.4.2
requests==2.31.0
sniffio==1.3.0
sqlparse==0.4.4
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.7
